terminal: move reading password from terminal here
